Add files via upload
Codsoft is a virtual internship program that provides students with hands-on experience through assigned tasks. I have started a one-month virtual internship with Codsoft as a July batch. This approach is helpful for knowledge acquisition.

Task 1: Build a model to predict whether a passenger on the Titanic survived or not.
Task 2: Build a model to predict the rating of a movie based on features like genre, director, and actors.
Task 3: Build a model to classify three different features of flowers.
Task 4: Predict sales according to given features.
Task 5: Build a machine learning model to identify fraudulent credit card transactions